+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: junyoung <junyoung@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/18 21:52:58 by junyoung          #+#    #+#             */
-/*   Updated: 2023/05/15 18:21:04 by junyoung         ###   ########seoul.kr  */
+/*   Updated: 2023/05/30 17:45:56 by junyoung         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,8 @@ typedef struct s_list
 	struct s_list	*next;
 }	t_list;
 
-int		ft_atoi(const char *str);
-int		ft_isdigit(int c);
-char	**ft_split(char const *s, char c);
+long long	ft_atoi(const char *str);
+int			ft_isdigit(int c);
+char		**ft_split(char const *s, char c);
 
 #endif
